Untrack stats save file + create stats save file if it doesnt exist

+++ tests/infra/test_txt_statistics_dao.py
@@ -7,7 +7,7 @@ from src.infra.txt_statistics_dao import TxtStatisticsDao
 
 
 class TxtStatisticsDaoTest(unittest.TestCase):
-    TEST_FILE = "tests/tmp.txt"
+    TEST_FILE = "tests/data/tmp.txt"
 
     def setUp(self):
         self.statistics_dao = TxtStatisticsDao(self.TEST_FILE)
@@ -21,7 +21,14 @@ class TxtStatisticsDaoTest(unittest.TestCase):
 
         loaded_statistics = self.statistics_dao.load()
 
-        self.assertEqual(self.statistics._stats, loaded_statistics._stats)
+        self.assertEqual(self.statistics, loaded_statistics)
+
+    def test_create_default_stats_if_save_file_doesnt_exist(self):
+        loaded_statistics = self.statistics_dao.load()
+
+        self.assertEqual(loaded_statistics, Statistics())
 
     def tearDown(self):
-        os.remove(self.TEST_FILE)
+        if os.path.exists(self.TEST_FILE):
+            os.remove(self.TEST_FILE)
+            os.rmdir(os.path.dirname(self.TEST_FILE))
